filer_Doc/requirements.txt
asgiref==3.2.7
Django==3.0.4
django-filer==1.7.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==2.1.2
djangorestframework==3.11.0
easy-thumbnails==2.7
Pillow==7.0.0
psycopg2-binary==2.8.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
Unidecode==1.1.1
